nex-kms/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nex-kms"
version = "0.0.1"
description = "Utilities for secrets in NEX python packages."
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "crcmod>=1.7",
    "google-cloud-kms>=2.21",
]

[project.entry-points.'nex-dev.subcommands']
kms = "nex_kms.cli:cli"

[tool.hatch.build]
exclude = [
    "*.dec"
]